+++ evaluation/performance/gpuFFT.cpp
@@ -21,14 +21,15 @@ struct EntryFFTCPU : public PerformanceTaskDefault {
     uint64 start = 0;
     cv::dft(complexI, tmp, cv::DFT_COMPLEX_OUTPUT);
     for (int i = 0; i != runs; ++i) {
-      start = cv::getTickCount();
+      start = static_cast<uint64>(cv::getTickCount());
       cv::dft(complexI, tmp, cv::DFT_COMPLEX_OUTPUT);
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount()) - start);
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((static_cast<uint64>(cv::getTickCount()) - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 
@@ -54,17 +55,18 @@ struct EntryFFTCL : public PerformanceTaskDefault {
     tmp.copyTo(tmp2);  // to RAM
 
     for (int i = 0; i != runs; ++i) {
-      start = cv::getTickCount();
+      start = static_cast<uint64>(cv::getTickCount());
       in = complexI.getUMat(cv::ACCESS_READ);  // to GPU
       cv::dft(in, tmp, cv::DFT_COMPLEX_OUTPUT);
       tmp.copyTo(tmp2);  // to RAM
 
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount()) - start);
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((static_cast<uint64>(cv::getTickCount()) - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 
@@ -88,15 +90,16 @@ struct EntryFFTCLNT : public PerformanceTaskDefault {
     uint64 start = 0;
     for (int i = 0; i != runs; ++i) {
       in = complexI.getUMat(cv::ACCESS_READ);  // to GPU
-      start = cv::getTickCount();
+      start = static_cast<uint64>(cv::getTickCount());
       cv::dft(in, tmp, cv::DFT_COMPLEX_OUTPUT);
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount()) - start);
       tmp.copyTo(tmp2);  // to RAM
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((static_cast<uint64>(cv::getTickCount()) - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 
@@ -122,16 +125,17 @@ struct EntryFFTCuda : public PerformanceTaskDefault {
     cv::cuda::dft(in, tmp, in.size());
     tmp.download(tmp2);
     for (int i = 0; i != runs; ++i) {
-      start = cv::getTickCount();
+      start = static_cast<uint64>(cv::getTickCount());
       in.upload(complexI);
       cv::cuda::dft(in, tmp, in.size());
       tmp.download(tmp2);
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount()) - start);
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((static_cast<uint64>(cv::getTickCount()) - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 
@@ -154,14 +158,15 @@ struct EntryFFTCudaNT : public PerformanceTaskDefault {
     in.upload(complexI);
     cv::cuda::dft(in, tmp, in.size());
     for (int i = 0; i != runs; ++i) {
-      start = cv::getTickCount();
+      start = static_cast<uint64>(cv::getTickCount());
       cv::cuda::dft(in, tmp, in.size());
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount()) - start);
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((static_cast<uint64>(cv::getTickCount()) - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 #endif
