+++ evaluation/performance/tests/threshold.cpp
@@ -36,14 +36,15 @@ struct Entry : public PerformanceTaskDefault {
     cv::Mat mag = filterRes["mag"].data;
     uint64 start = 0;
     for (int i = 0; i != loops; ++i) {
-      start = cv::getTickCount();
+      start = static_cast<uint64>(cv::getTickCount());
       thresholdRes = threshold->process(mag);
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount()) - start);
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (loops * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((static_cast<uint64>(cv::getTickCount()) - start) * 1000) /
+                       (loops * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 
   void saveResults(bool verbose) {
